Add some prints
Seems that hit_end is used wrongly.
It does not say if there will be no new matches, it specifies if for the
CURRENT match, adding more input could change it. But it does not apply
to new matches.

Still I'm not sure why these regexes change if hit_end is reported.